+++ CHANGES.md
@@ -41,6 +41,7 @@
 - Fix bug in keeping of table references when setting to keep all named references
 - Fix so that works when there are apostrophes in the path [#12](https://github.com/tamc/excel_to_code/issues/12)
 - Improve bracket elimination when simplifying arithmetic
+- Fix IF functions where comparison is an array
 
 ## Refactors